+++ gradle/libs.versions.toml
@@ -12,6 +12,7 @@ tck-common-testplan-gen = { module = "org.eclipse.dataspacetck:test-plan-generat
 junit-jupiter = { module = "org.junit.jupiter:junit-jupiter", version.ref = "jupiter" }
 junit-platform-engine = { module = "org.junit.platform:junit-platform-suite-engine", version.ref = "junit" }
 plugin-docker-remoteapi = { module = "com.bmuschko.docker-remote-api:com.bmuschko.docker-remote-api.gradle.plugin", version = "9.4.0" }
+plugin-nexus-publish = { module = "io.github.gradle-nexus:publish-plugin", version = "2.0.0" }
 
 [plugins]
 nexuspublishing = { id = "io.github.gradle-nexus.publish-plugin", version = "2.0.0" }
